Consolidate JS into a single file
