--- compiler/core/src/component.bs.js
@@ -1,46 +1,0 @@
-// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
-'use strict';
-
-var Block                   = require("bs-platform/lib/js/block.js");
-var Ast$LonaCompilerCore    = require("./ast.bs.js");
-var Layer$LonaCompilerCore  = require("./layer.bs.js");
-var Logic$LonaCompilerCore  = require("./logic.bs.js");
-var Decode$LonaCompilerCore = require("./decode.bs.js");
-
-function generate(name, json) {
-  var rootLayer = Decode$LonaCompilerCore.Component[/* rootLayer */1](json);
-  var logic = Logic$LonaCompilerCore.addVariableDeclarations(Decode$LonaCompilerCore.Component[/* logic */2](json));
-  var assignments = Layer$LonaCompilerCore.parameterAssignmentsFromLogic(rootLayer, logic);
-  var rootLayerAST = Layer$LonaCompilerCore.toJavaScriptAST(assignments, rootLayer);
-  var styleSheetAST = Layer$LonaCompilerCore.toJavaScriptStyleSheetAST(rootLayer);
-  var logicAST = Ast$LonaCompilerCore.JavaScript[/* optimize */3](Logic$LonaCompilerCore.toJavaScriptAST(logic));
-  return Ast$LonaCompilerCore.JavaScript[/* prepareForRender */4](/* Program */Block.__(16, [/* :: */[
-                  /* Class */Block.__(3, [
-                      name,
-                      /* Some */["React.Component"],
-                      /* :: */[
-                        /* Method */Block.__(4, [
-                            "render",
-                            /* [] */0,
-                            /* :: */[
-                              logicAST,
-                              /* :: */[
-                                /* Return */Block.__(0, [rootLayerAST]),
-                                /* [] */0
-                              ]
-                            ]
-                          ]),
-                        /* [] */0
-                      ]
-                    ]),
-                  /* :: */[
-                    styleSheetAST,
-                    /* [] */0
-                  ]
-                ]]));
-}
-
-var JavaScript = /* module */[/* generate */generate];
-
-exports.JavaScript = JavaScript;
-/* Layer-LonaCompilerCore Not a pure module */
