--- compiler/core/src/component.re
@@ -1,22 +1,0 @@
-module JavaScript = {
-  let generate = (name, json) => {
-    let rootLayer = json |> Decode.Component.rootLayer;
-    let logic = json |> Decode.Component.logic |> Logic.addVariableDeclarations;
-    let assignments = Layer.parameterAssignmentsFromLogic(rootLayer, logic);
-    let rootLayerAST = rootLayer |> Layer.toJavaScriptAST(assignments);
-    let styleSheetAST = rootLayer |> Layer.toJavaScriptStyleSheetAST;
-    let logicAST = logic |> Logic.toJavaScriptAST |> Ast.JavaScript.optimize;
-    Ast.JavaScript.(
-      Program([
-        Class(
-          name,
-          Some("React.Component"),
-          [Method("render", [], [logicAST, Return(rootLayerAST)])]
-        ),
-        styleSheetAST
-      ])
-    )
-    /* Renames variables */
-    |> Ast.JavaScript.prepareForRender
-  };
-};
