+++ compiler/core/src/layer.re
@@ -197,65 +197,3 @@ let layerTypeToString = (x) =>
   };
 
 let mapBindings = (f, map) => map |> StringMap.bindings |> List.map(f);
-
-let createStyleAttributeAST = (layerName, styles) =>
-  Ast.JavaScript.(
-    JSXAttribute(
-      "style",
-      ArrayLiteral([
-        Identifier(["styles", layerName]),
-        ObjectLiteral(
-          styles
-          |> mapBindings(
-               ((key, value)) =>
-                 ObjectProperty(Identifier([key]), Logic.logicValueToJavaScriptAST(value))
-             )
-        )
-      ])
-    )
-  );
-
-let rec toJavaScriptAST = (variableMap, layer: Types.layer) => {
-  open Ast.JavaScript;
-  let (_, mainParams) = layer.parameters |> parameterMapToLogicValueMap |> splitParamsMap;
-  let (styleVariables, mainVariables) =
-    (
-      switch (LayerMap.find_opt(layer, variableMap)) {
-      | Some(map) => map
-      | None => StringMap.empty
-      }
-    )
-    |> splitParamsMap;
-  let main = StringMap.assign(mainParams, mainVariables);
-  let styleAttribute = createStyleAttributeAST(layer.name, styleVariables);
-  let attributes =
-    main
-    |> mapBindings(((key, value)) => JSXAttribute(key, Logic.logicValueToJavaScriptAST(value)));
-  JSXElement(
-    layerTypeToString(layer.typeName),
-    [styleAttribute, ...attributes],
-    layer.children |> List.map(toJavaScriptAST(variableMap))
-  )
-};
-
-let toJavaScriptStyleSheetAST = (layer: Types.layer) => {
-  open Ast.JavaScript;
-  let createStyleObjectForLayer = (layer: Types.layer) => {
-    let styleParams = layer.parameters |> StringMap.filter((key, _) => parameterIsStyle(key));
-    ObjectProperty(
-      Identifier([layer.name]),
-      ObjectLiteral(
-        styleParams
-        |> StringMap.bindings
-        |> List.map(((key, value)) => ObjectProperty(Identifier([key]), Literal(value)))
-      )
-    )
-  };
-  let styleObjects = layer |> flatten |> List.map(createStyleObjectForLayer);
-  VariableDeclaration(
-    AssignmentExpression(
-      Identifier(["styles"]),
-      CallExpression(Identifier(["StyleSheet", "create"]), [ObjectLiteral(styleObjects)])
-    )
-  )
-};
