+++ compiler/core/src/logic.re
@@ -101,52 +101,3 @@ let addVariableDeclarations = (node) => {
        node
      )
 };
-
-let logicValueToJavaScriptAST = (x) =>
-  switch x {
-  | Identifier(_, path) => Ast.JavaScript.Identifier(path)
-  | Literal(x) => Literal(x)
-  | None => Unknown
-  };
-
-let rec toJavaScriptAST = (node) => {
-  let fromCmp = (x) =>
-    switch x {
-    | Types.Eq => Ast.JavaScript.Eq
-    | Neq => Neq
-    | Gt => Gt
-    | Gte => Gte
-    | Lt => Lt
-    | Lte => Lte
-    | Unknown => Noop
-    };
-  switch node {
-  | Assign(a, b) =>
-    Ast.JavaScript.AssignmentExpression(logicValueToJavaScriptAST(b), logicValueToJavaScriptAST(a))
-  | IfExists(a, body) =>
-    ConditionalStatement(logicValueToJavaScriptAST(a), [toJavaScriptAST(body)])
-  | Block(body) => Ast.JavaScript.Block(body |> List.map(toJavaScriptAST))
-  | If(a, cmp, b, body) =>
-    let condition =
-      Ast.JavaScript.BooleanExpression(
-        logicValueToJavaScriptAST(a),
-        fromCmp(cmp),
-        logicValueToJavaScriptAST(b)
-      );
-    ConditionalStatement(condition, [toJavaScriptAST(body)])
-  | Add(lhs, rhs, value) =>
-    let addition =
-      Ast.JavaScript.BooleanExpression(
-        logicValueToJavaScriptAST(lhs),
-        Plus,
-        logicValueToJavaScriptAST(rhs)
-      );
-    AssignmentExpression(logicValueToJavaScriptAST(value), addition)
-  | Let(value) =>
-    switch value {
-    | Identifier(_, path) => Ast.JavaScript.VariableDeclaration(Ast.JavaScript.Identifier(path))
-    | _ => Unknown
-    }
-  | None => Unknown
-  }
-};
